# api/openapi.yaml
openapi: 3.0.3
info:
  title: WhereToLive HTTP API
  version: 0.1.0
  description: Authoritative HTTP contract for the WhereToLive modular monolith.
servers:
  - url: /api
paths:
...
  /public/places:
    get:
      operationId: searchPublicPlaces
      summary: Search published places by stable slug, name or multilingual alias
      tags: [public-places]
      security: []
      parameters:
        - name: q
          in: query
          schema: {type: string, maxLength: 120, default: ""}
        - $ref: "#/components/parameters/PublicLocale"
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 50, default: 20}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 10000, default: 0}
      responses:
        "200":
          description: Published places only; no session is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPlacePage"
        "400":
          $ref: "#/components/responses/PublicInvalidQuery"
        "503":
          $ref: "#/components/responses/PublicUnavailable"
        "500":
          $ref: "#/components/responses/PublicInternalError"
  /public/places/{slug}:
    get:
      operationId: getPublicPlace
      summary: Read a published place by stable language-independent slug
      tags: [public-places]
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: {type: string, maxLength: 120, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'}
        - $ref: "#/components/parameters/PublicLocale"
      responses:
        "200":
          description: Published place with requested display language or canonical fallback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPlace"
        "400":
          $ref: "#/components/responses/PublicInvalidQuery"
        "404":
          description: Not found; unpublished places are indistinguishable from missing ones
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          $ref: "#/components/responses/PublicUnavailable"
        "500":
          $ref: "#/components/responses/PublicInternalError"
...
  /platform/auth/login:
    post:
      operationId: platformLogin
      summary: Establish a global platform-administrator session
      tags: [platform-authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformLoginRequest"
      responses:
        "200":
          description: Platform session established with a distinct refresh cookie
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
  /platform/auth/refresh:
    post:
      operationId: platformRefresh
      summary: Rotate a platform-administrator refresh secret
      tags: [platform-authentication]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Platform session rotated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
  /platform/auth/logout:
    post:
      operationId: platformLogout
      summary: Revoke the current platform-administrator session
      tags: [platform-authentication]
      security:
        - platformBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204":
          description: Platform session revoked and cookies cleared
          headers:
            Set-Cookie:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
  /platform/tenants:
    post:
      operationId: provisionPlatformTenant
      summary: Atomically provision a tenant as a global platform administrator
      tags: [platform-tenants]
      security:
        - platformBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionTenantRequest"
      responses:
        "201":
          description: Tenant, first administrator, organization, and role were created atomically
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisionTenantResponse"
        "200":
          description: The same completed provisioning request was returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisionTenantResponse"
        "400":
          description: Invalid provisioning request
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
        "409":
          description: Idempotency key was reused with different input
    get:
      operationId: listPlatformTenants
      summary: List tenants as a global platform administrator
      tags: [platform-tenants]
      security:
        - platformBearerAuth: []
      responses:
        "200":
          description: Tenant summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformTenant"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
  /platform/tenants/{tenantId}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    patch:
      operationId: updatePlatformTenant
      summary: Update a tenant's mutable platform profile
      tags: [platform-tenants]
      security:
        - platformBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlatformTenantRequest"
      responses:
        "204":
          description: Tenant profile and audit evidence committed
        "400":
          description: Invalid tenant profile update
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
        "404":
          description: Tenant not found
        "409":
          description: Tenant profile was changed by another administrator
  /platform/tenants/{tenantId}/suspend:
    post:
      operationId: suspendPlatformTenant
      summary: Suspend an active tenant with auditable reason
      tags: [platform-tenants]
      security:
        - platformBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantLifecycleRequest"
      responses:
        "204":
          description: Tenant suspended and audit evidence committed
        "400":
          description: Invalid lifecycle transition
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
        "404":
          description: Tenant not found
  /platform/tenants/{tenantId}/reactivate:
    post:
      operationId: reactivatePlatformTenant
      summary: Reactivate a suspended tenant with auditable reason
      tags: [platform-tenants]
      security:
        - platformBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantLifecycleRequest"
      responses:
        "204":
          description: Tenant reactivated and audit evidence committed
        "400":
          description: Invalid lifecycle transition
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
        "404":
          description: Tenant not found
...
  /auth/refresh:
    post:
      operationId: refresh
      summary: Rotate the refresh secret and issue an access token
      tags: [authentication]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Session rotated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
...
components:
  securitySchemes:
    consumerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    platformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PublicLocale:
      name: locale
      in: query
      schema:
        type: string
        enum: [en, zh-CN, de, fr, es]
        default: en
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 256
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    PublicInvalidQuery:
      description: Invalid query
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PublicUnavailable:
      description: Catalogue is unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PublicInternalError:
      description: Internal error without private details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    AuthenticationFailed:
      description: Tenant or credentials are invalid, or the session is unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    CsrfFailed:
      description: CSRF validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    PublicPlaceType:
      type: string
      enum: [country, region, island, city, district]
...
    PublicPlace:
      type: object
      required: [id, slug, name, displayName, type, parentId, countryCode, timezone, latitude, longitude, currency, languages, coverageLevel, publishedAt]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string, maxLength: 120, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'}
        name: {type: string, description: Canonical name}
        displayName: {type: string, description: Preferred localized alias or canonical name}
        type: {$ref: "#/components/schemas/PublicPlaceType"}
        parentId: {type: string, format: uuid, nullable: true}
        countryCode: {type: string, pattern: '^[A-Z]{2}$'}
        timezone: {type: string, nullable: true}
        latitude: {type: number, format: double, minimum: -90, maximum: 90, nullable: true}
        longitude: {type: number, format: double, minimum: -180, maximum: 180, nullable: true}
        currency: {type: string, pattern: '^[A-Z]{3}$', nullable: true}
        languages:
          type: array
          items: {type: string}
        coverageLevel: {type: integer, minimum: 1, maximum: 3}
        publishedAt: {type: string, format: date-time, description: Catalogue publication time; not fact verification time}
    PublicPlacePage:
      type: object
      required: [items, nextOffset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicPlace"
        nextOffset: {type: integer, nullable: true, minimum: 1}
...
    PlatformLoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 254
        password:
          type: string
          minLength: 1
          maxLength: 1024
          writeOnly: true
    PlatformTenant:
      type: object
      additionalProperties: false
      required: [id, slug, displayName, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        displayName:
          type: string
        status:
          type: string
          enum: [provisioning, active, suspended]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdatePlatformTenantRequest:
      type: object
      additionalProperties: false
      required: [displayName, expectedUpdatedAt, reason]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 128
        expectedUpdatedAt:
          type: string
          format: date-time
        reason:
          type: string
          minLength: 1
          maxLength: 512
    ProvisionTenantRequest:
      type: object
      additionalProperties: false
      required: [slug, displayName, rootDepartmentName, administratorUsername, reason]
      properties:
        slug:
          type: string
          minLength: 1
          maxLength: 63
          pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$"
        displayName:
          type: string
          minLength: 1
          maxLength: 128
        rootDepartmentName:
          type: string
          minLength: 1
          maxLength: 128
        administratorUsername:
          type: string
          minLength: 1
          maxLength: 128
        administratorEmail:
          type: string
          format: email
          maxLength: 254
        reason:
          type: string
          minLength: 1
          maxLength: 500
    ProvisionTenantResponse:
      type: object
      additionalProperties: false
      required: [tenantId, created]
      properties:
        tenantId:
          type: string
          format: uuid
        created:
          type: boolean
    TenantLifecycleRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
...
    AccessTokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken, tokenType, expiresIn, csrfToken]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          format: int64
          minimum: 1
        csrfToken:
          type: string
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        traceId:
          type: string